+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - OpenAI: New `responses_store` model argument to control whether the `store` option is enabled (it is enabled by default for reasoning models to support reasoning playback).
 - OpenAI: Support for [flex processing](https://inspect.aisi.org.uk/providers.html#flex-processing), which provides lower inference costs in exchange for slower response times and occasional resource unavailability (added in v1.75.0, which is now required).
 - Google: Support for `reasoning_tokens` option for Gemini 2.5 models.
+- Grok: Support for `reasoning_effort` option and capturing reasoning content.
 - Human Agent: Added `user` parameter for running the human agent cli as a given user.
 - Sandbox API: Added `user` parameter to `connection()` method for getting connection details for a given user.
 - Docker: Support sample metadata interpolation for image names in compose files.